+++ docSite/content/docs/development/upgrading/48.md
@@ -22,4 +22,5 @@ FastGPT workflow V2上线，支持更加简洁的工作流模式。
 3. 新增 - 定时执行应用。可轻松实现定时任务。
 4. 新增 - 插件自定义输入优化，可以渲染输入组件。
 5. 优化 - 工作流连线，可以四向连接，方便构建循环工作流。
-6. 优化 - worker进程管理，并将计算 Token 任务分配给 worker 进程。
+6. 优化 - 工作流上下文传递，性能🚀。
+7. 优化 - worker进程管理，并将计算 Token 任务分配给 worker 进程。
