+++ projects/app/package.json
@@ -23,6 +23,7 @@
     "@fastgpt/service": "workspace:*",
     "@fastgpt/web": "workspace:*",
     "@fortaine/fetch-event-source": "^3.0.6",
+    "@node-rs/jieba": "1.10.0",
     "@tanstack/react-query": "^4.24.10",
     "@types/nprogress": "^0.2.0",
     "ahooks": "^3.7.11",
@@ -43,6 +44,7 @@
     "nanoid": "^4.0.1",
     "next": "13.5.2",
     "next-i18next": "15.2.0",
+    "nextjs-node-loader": "^1.1.5",
     "nprogress": "^0.2.0",
     "react": "18.2.0",
     "react-day-picker": "^8.7.1",
@@ -58,9 +60,8 @@
     "remark-math": "^5.1.1",
     "request-ip": "^3.3.0",
     "sass": "^1.58.3",
-    "zustand": "^4.3.5",
-    "nextjs-node-loader": "^1.1.5",
-    "@node-rs/jieba": "1.10.0"
+    "use-context-selector": "^1.4.4",
+    "zustand": "^4.3.5"
   },
   "devDependencies": {
     "@svgr/webpack": "^6.5.1",
