+++ projects/app/src/components/core/workflow/Flow/ImportSettings.tsx
@@ -3,7 +3,8 @@ import { Textarea, Button, ModalBody, ModalFooter } from '@chakra-ui/react';
 import MyModal from '@fastgpt/web/components/common/MyModal';
 import { useTranslation } from 'next-i18next';
 import { useToast } from '@fastgpt/web/hooks/useToast';
-import { useFlowProviderStore } from './FlowProvider';
+import { useContextSelector } from 'use-context-selector';
+import { WorkflowContext } from '../context';
 
 type Props = {
   onClose: () => void;
@@ -12,7 +13,8 @@ type Props = {
 const ImportSettings = ({ onClose }: Props) => {
   const { t } = useTranslation();
   const { toast } = useToast();
-  const { setNodes, setEdges, initData } = useFlowProviderStore();
+
+  const initData = useContextSelector(WorkflowContext, (v) => v.initData);
   const [value, setValue] = useState('');
 
   return (
