+++ projects/app/src/components/core/workflow/Flow/nodes/render/RenderInput/templates/NumberInput.tsx
@@ -7,10 +7,11 @@ import {
   NumberInputField,
   NumberInputStepper
 } from '@chakra-ui/react';
-import { useFlowProviderStore } from '../../../../FlowProvider';
+import { useContextSelector } from 'use-context-selector';
+import { WorkflowContext } from '@/components/core/workflow/context';
 
 const NumberInputRender = ({ item, nodeId }: RenderInputProps) => {
-  const { onChangeNode } = useFlowProviderStore();
+  const onChangeNode = useContextSelector(WorkflowContext, (v) => v.onChangeNode);
 
   const Render = useMemo(() => {
     return (
